Add delete and complete features

+++ src/App.jsx
@@ -9,11 +9,14 @@ import DataContext from './Context/Contextapi'
 const App = () => {
 
   const [todoArray,setTodoArray] = useState([]);
+  const [activeTab,setActiveTab] = useState('All');
+  // const [binItems,setBinItems] = useState([]);
 
   const fetchData = async() => {
 
     const data = await api.get("/todolist")
-    setTodoArray(data.data)
+
+    setTodoArray(data.data.filter((task) => task.moveToTrash === false && task.complete === false))
     
   }
   
@@ -21,7 +24,6 @@ const App = () => {
     fetchData()
     
   },[])
-  console.log(todoArray);
 
   const addNewTodoToServer = async(newTask) => {
     await api.post("/todoList",newTask)
@@ -31,32 +33,94 @@ const App = () => {
 
   // delete task from the api and the state
   const deleteTodo = async(id) => {
-    console.log(typeof id);
-    await api.delete(`/todolist/${id}`)
-    const newTodo = [...todoArray];
-    setTodoArray(newTodo.filter((task)=> task.id !== id));
+    if(confirm ('You sure!This task will be remove permently!')){
+      await api.delete(`/todolist/${id}`)
+      const newTodo = [...todoArray];
+      setTodoArray(newTodo.filter((task)=> task.id !== id));
+      return;
+    }
+    return;
+  }
+
+  const moveToTrash = async(id,currentStat) => {
+
+    await api.patch(`/todoList/${id}`,{moveToTrash:!currentStat});
+  
+    const currentTodo = todoArray.map((task) => {
+      if(task.id === id){
+        task.moveToTrash = !currentStat;
+        return task
+      }
+      return task;
+    });
+
+    
+    setTodoArray(currentTodo.filter((task) => task.moveToTrash === false))
 
   }
 
   // partial upadte the data 
   const HandleComplete = async(id,currentStat) => {
 
-    await api.patch(`/todoList/${id}`,{complete:!currentStat});
-  
-    const currentTodo = todo.filter((task) => {
-      if(task.id === id){
-        task.complete = !currentStat;
-        return task
-      }
-      return todo;
-    });
-    
-    setTodo([...currentTodo])
 
+    await api.patch(`/todoList/${id}`,{complete:!currentStat});
+
+    if(!currentStat){
+
+      if(confirm ("This task will be move to 'Done' section.")){
+        
+        const currentTodo = todoArray.map((task) => {
+          if(task.id === id){
+            task.complete = !currentStat;
+            return task
+          }
+          return task;
+        });
+        
+        setTodoArray(currentTodo.filter((task) => task.complete === false))
+      }
+    }else {
+
+      if(confirm ("This task will set to undone.")){
+        
+        const currentTodo = todoArray.map((task) => {
+          if(task.id === id){
+            task.complete = !currentStat;
+            return task
+          }
+          return task;
+        });
+        
+        setTodoArray(currentTodo.filter((task) => task.complete === true))
+      }
+    }
+      
   }
 
+  // filter by categories
+  const renderByType = async(type) => {
+    const data = await api.get("/todolist")
+    if(type === "All"){
+
+      setTodoArray(data.data.filter((task) => task.moveToTrash !== true && task.complete === false))
+      
+    }else if(type === "Trash bin") {
+      setTodoArray(data.data.filter((task) => task.moveToTrash === true ))
+    }else if(type === "Done") {
+      setTodoArray(data.data.filter((task) => task.complete === true ))
+    }
+    else {
+      const currentTodoArray = [...data.data];
+      const filteredArray = currentTodoArray.filter((task) => task.type === type && task.moveToTrash === false && task.complete === false
+      )
+      setTodoArray(filteredArray);
+    }
+
+  }
+  
+
   return (
-    <DataContext.Provider value={{todoArray,addNewTodoToServer,deleteTodo }}>
+    <DataContext.Provider value={{todoArray,addNewTodoToServer,deleteTodo, HandleComplete, renderByType, activeTab,setActiveTab, moveToTrash }}>
       
       <header>
         <Headbar />
@@ -73,6 +137,7 @@ const App = () => {
       <div>
         <Card />
       </div>
+
 
     </DataContext.Provider>
   )
